+++ src/main/java/com/group5/MiniSurveyMonkey/Login/DBUserDetailsService.java
@@ -18,7 +18,7 @@ public class DBUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        LocalUser user = userRepository.findByUser(username);
+        User user = userRepository.findByUser(username);
         if (user == null) throw new UsernameNotFoundException("User not found");
 
         List<SimpleGrantedAuthority> authorities = Arrays.asList(new SimpleGrantedAuthority("USER"));
