--- src/main/java/com/group5/MiniSurveyMonkey/Login/LocalUser.java
@@ -1,57 +1,0 @@
-package com.group5.MiniSurveyMonkey.Login;
-
-import javax.persistence.*;
-
-@Entity(name = "User")
-@Table(name = "user")
-public class LocalUser {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String user;
-    private String password;
-    private boolean active;
-    private String roles;
-
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getUser() {
-        return user;
-    }
-
-    public void setUser(String user) {
-        this.user = user;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public boolean isActive() {
-        return active;
-    }
-
-    public void setActive(boolean active) {
-        this.active = active;
-    }
-
-    public String getRoles() {
-        return roles;
-    }
-
-    public void setRoles(String roles) {
-        this.roles = roles;
-    }
-
-}
